backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Mail
Flask-CORS
python-dotenv
bcrypt
